+++ app/controllers/application_controller.rb
@@ -1,8 +1,7 @@
 class ApplicationController < ActionController::Base
-  # before_action :authenticate_user!
 
   def default_url_options
-  { host: ENV["ethikly"] || "localhost:3000" }
+  { host: ENV["DOMAIN"] || "localhost:3000" }
   end
 
   def change_avatar
